# Use more cache friendly implementations