--- a/expr/seriesaggregators.go
+++ b/expr/seriesaggregators.go
@@ -66,96 +66,72 @@ func crossSeriesAvg(in []models.Series, out *[]schema.Point) {
 
 func crossSeriesMin(in []models.Series, out *[]schema.Point) {
 	for i := 0; i < len(in[0].Datapoints); i++ {
-		nan := true
-		min := math.Inf(1)
-		for j := 0; j < len(in); j++ {
-			p := in[j].Datapoints[i].Val
-			if !math.IsNaN(p) && p < min {
-				nan = false
-				min = p
+		*out = append(*out, in[0].Datapoints[i])
+	}
+
+	for i := 1; i < len(in); i++ {
+		dps := in[i].Datapoints
+		for j := 0; j < len(in[i].Datapoints); j++ {
+			p := dps[j].Val
+			if !math.IsNaN(p) {
+				v := (*out)[j].Val
+				if math.IsNaN(v) || v > p {
+					(*out)[j].Val = p
+				}
 			}
 		}
-
-		point := schema.Point{
-			Ts: in[0].Datapoints[i].Ts,
-		}
-		if nan {
-			point.Val = math.NaN()
-		} else {
-			point.Val = min
-		}
-
-		*out = append(*out, point)
 	}
 }
 
 func crossSeriesMax(in []models.Series, out *[]schema.Point) {
 	for i := 0; i < len(in[0].Datapoints); i++ {
-		nan := true
-		max := math.Inf(-1)
-		for j := 0; j < len(in); j++ {
-			p := in[j].Datapoints[i].Val
-			if !math.IsNaN(p) && p > max {
-				nan = false
-				max = p
+		*out = append(*out, in[0].Datapoints[i])
+	}
+
+	for i := 1; i < len(in); i++ {
+		dps := in[i].Datapoints
+		for j := 0; j < len(in[i].Datapoints); j++ {
+			p := dps[j].Val
+			if !math.IsNaN(p) {
+				v := (*out)[j].Val
+				if math.IsNaN(v) || v < p {
+					(*out)[j].Val = p
+				}
 			}
 		}
-
-		point := schema.Point{
-			Ts: in[0].Datapoints[i].Ts,
-		}
-		if nan {
-			point.Val = math.NaN()
-		} else {
-			point.Val = max
-		}
-
-		*out = append(*out, point)
 	}
 }
 
 func crossSeriesSum(in []models.Series, out *[]schema.Point) {
 	for i := 0; i < len(in[0].Datapoints); i++ {
-		nan := true
-		sum := float64(0)
-		for j := 0; j < len(in); j++ {
-			p := in[j].Datapoints[i].Val
+		*out = append(*out, in[0].Datapoints[i])
+	}
+
+	for i := 1; i < len(in); i++ {
+		dps := in[i].Datapoints
+		for j := 0; j < len(in[i].Datapoints); j++ {
+			p := dps[j].Val
 			if !math.IsNaN(p) {
-				nan = false
-				sum += p
+				if math.IsNaN((*out)[j].Val) {
+					(*out)[j].Val = p
+				} else {
+					(*out)[j].Val += p
+				}
 			}
 		}
-		point := schema.Point{
-			Ts: in[0].Datapoints[i].Ts,
-		}
-		if nan {
-			point.Val = math.NaN()
-		} else {
-			point.Val = sum
-		}
-
-		*out = append(*out, point)
 	}
 }
 
 func crossSeriesMultiply(in []models.Series, out *[]schema.Point) {
 	for i := 0; i < len(in[0].Datapoints); i++ {
-		product := float64(1)
-		for j := 0; j < len(in); j++ {
-			p := in[j].Datapoints[i].Val
-			if math.IsNaN(p) {
-				// NaN * anything equals NaN()
-				product = math.NaN()
-				break
-			}
-			product *= p
-		}
-		point := schema.Point{
-			Ts:  in[0].Datapoints[i].Ts,
-			Val: product,
-		}
+		*out = append(*out, in[0].Datapoints[i])
+	}
 
-		*out = append(*out, point)
+	for i := 1; i < len(in); i++ {
+		dps := in[i].Datapoints
+		for j := 0; j < len(in[i].Datapoints); j++ {
+			(*out)[j].Val *= dps[j].Val
+		}
 	}
 }
 
@@ -190,68 +166,52 @@ func crossSeriesMedian(in []models.Series, out *[]schema.Point) {
 
 func crossSeriesDiff(in []models.Series, out *[]schema.Point) {
 	for i := 0; i < len(in[0].Datapoints); i++ {
-		val := math.NaN()
-		for j := 0; j < len(in); j++ {
-			p := in[j].Datapoints[i].Val
+		*out = append(*out, in[0].Datapoints[i])
+	}
+
+	for i := 1; i < len(in); i++ {
+		for j := 0; j < len(in[i].Datapoints); j++ {
+			p := in[i].Datapoints[j].Val
 			if !math.IsNaN(p) {
-				if math.IsNaN(val) {
-					val = p
+				if math.IsNaN((*out)[j].Val) {
+					(*out)[j].Val = p
 				} else {
-					val -= p
+					(*out)[j].Val -= p
 				}
 			}
-		}
-		point := schema.Point{
-			Ts:  in[0].Datapoints[i].Ts,
-			Val: val,
-		}
 
-		*out = append(*out, point)
+		}
 	}
 }
 
 func crossSeriesStddev(in []models.Series, out *[]schema.Point) {
-	averages := make([]schema.Point, 0, len(in[0].Datapoints))
-	crossSeriesAvg(in, &averages)
+	crossSeriesAvg(in, out)
 
 	for i := 0; i < len(in[0].Datapoints); i++ {
-		point := schema.Point{
-			Ts:  in[0].Datapoints[i].Ts,
-			Val: math.NaN(),
-		}
-
-		if !math.IsNaN(averages[i].Val) {
+		if !math.IsNaN((*out)[i].Val) {
 			num := float64(0)
 			totalDeviationSquared := float64(0)
 			for j := 0; j < len(in); j++ {
 				p := in[j].Datapoints[i].Val
 				if !math.IsNaN(p) {
 					num++
-					deviation := p - averages[i].Val
+					deviation := p - (*out)[i].Val
 					totalDeviationSquared += deviation * deviation
 				}
 			}
 
-			point.Val = math.Sqrt(totalDeviationSquared / num)
+			(*out)[i].Val = math.Sqrt(totalDeviationSquared / num)
 		}
-
-		*out = append(*out, point)
 	}
 }
 
 func crossSeriesRange(in []models.Series, out *[]schema.Point) {
-	maxes := make([]schema.Point, 0, len(in[0].Datapoints))
 	mins := make([]schema.Point, 0, len(in[0].Datapoints))
 
-	crossSeriesMax(in, &maxes)
+	crossSeriesMax(in, out)
 	crossSeriesMin(in, &mins)
 
 	for i := 0; i < len(in[0].Datapoints); i++ {
-		point := schema.Point{
-			Ts:  in[0].Datapoints[i].Ts,
-			Val: maxes[i].Val - mins[i].Val,
-		}
-
-		*out = append(*out, point)
+		(*out)[i].Val -= mins[i].Val
 	}
 }