@Dieterbe 2018-12-20
So in master, we iterate, for each point in time, over all series and the output series.
In this PR, we iterate, for each input series, over all points for the input series and the output series

It makes sense that the latter could be faster, though it's a suprise to me why a profile would show time spent in math.IsNaN() though. did you mean time is spent on the `math.IsNaN(p)` line, or within math.IsNaN() itself? i would expect the former, not the latter.